ecom-webapp-backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
numpy==1.26.4
pandas==2.2.1
python-dotenv==1.0.1
requests==2.31.0
blosum==2.0.2
fastjsonschema==2.19.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
nltk==3.8.1
openai==1.17.1
python-dotenv==1.0.1
python-json-logger==2.0.7
tensorflow==2.16.1
textblob==0.18.0.post0
scipy==1.12.0
scikit-learn==1.4.1.post1
Werkzeug==3.0.2
gunicorn==20.1.0
wasabi==1.1.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
wheel==0.42.0
wrapt==1.16.0
matplotlib==3.8.3